Change class name to make it more meaning
